pr feedback: u64 -> u32 + revert

+++ ledger/src/blockstore_meta.rs
@@ -143,7 +143,7 @@ pub struct MerkleRootMeta {
     /// The merkle root
     merkle_root: Hash,
     /// The first received shred index
-    first_received_shred_index: u64,
+    first_received_shred_index: u32,
     /// The shred type of the first received shred
     first_received_shred_type: ShredType,
 }
